requirements.txt
django==5.2.*
dj-database-url
python-dotenv
rasterio
psycopg2-binary
earthengine-api
